better way to conditionally compile the cuda wrapper
