+++ R/utils.R
@@ -26,3 +26,7 @@ add_class_definition <- function(r6_class_generator){
   r6_class_generator$lock()
   r6_class_generator
 }
+
+torch_gc_wrapper <- function() {
+  gc(full = FALSE)
+}
