+++ lantern/src/AllocatorUtils.h
@@ -1,0 +1,3 @@
+#pragma once
+
+extern void (*call_r_gc) ();
