+++ src/allocator.cpp
@@ -3,8 +3,9 @@
 
 void call_r_gc ()
 {
-  static Rcpp::Function r_gc("gc");
-  r_gc(Rcpp::Named("full")=false);
+  static Rcpp::Environment torch("package:torch");
+  static Rcpp::Function gc = torch["torch_gc_wrapper"];
+  gc();
 }
 
 // [[Rcpp::export]]
